SCRIPTED_WORKFLOW/src/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.8.0
argparse>=1.4.0
pyyaml>=6.0.2
xarray>=2024.1.2
#xarray>=2025.1.2
glob2>=0.7 